[FEATURE] Automatic creation of preset thumbnails
To improve response times of the backend requests, thumbnails presets
are created automatically when new images are uploaded / created.

The configuration setting ``TYPO3.Media.autoGenerateThumbnailPresets`` can be used
to disable the feature.

If the configuration setting ``TYPO3.Media.asyncThumbnails`` is set to ``false`` the
created thumbnails are generated immediately, instead of upon request.

Resolves: NEOS-1671